+++ api/src/controllers/userController.ts
@@ -1,17 +1,20 @@
 
-import { BadRequestError, NotFoundError } from '../utils/errors';
+import { BadRequestError, ForbiddenError, NotFoundError } from '../utils/errors';
 import * as UserService from '../services/userService';
+import { logger } from '../utils/logger';
 
 export async function getUser(req: any, res: any, next: any) {
+    const id = Number(req.params?.id);
     try {
-        const id = Number(req.params.id);
-
         if (isNaN(id)) throw new BadRequestError('/id', 'id must be a number');
 
         const user = await UserService.getUser(req.params.id);
         if (!user) throw new NotFoundError(`Could not find user with id ${id}`);
-        res.status(200).json(user);
+        
+        logger.info("getUser: success. userId: " + user.id);
+        res.status(200).json(user.getBasic());
     } catch (error: any) {
+        logger.error("getUser: error with status " + error.status + ". userId: " + id);
         next(error);
     }
 }
@@ -19,34 +22,18 @@ export async function getUser(req: any, res: any, next: any) {
 export async function listUsers(req: any, res: any, next: any) {
     try {
         const users = await UserService.listUsers();
-        return res.status(200).json(users);
+
+        logger.info("listUsers: success. Users found: " + users.length);
+        return res.status(200).json(users.map(user => user.getBasic()));
     } catch (error: any) {
-        next(error);
-    }
-}
-
-export async function createUser(req: any, res: any, next: any) {
-    try {
-        const email = req.body?.email;
-        const username = req.body?.username;
-
-        if (!email) throw new BadRequestError('/body/email', 'missing required field');
-        if (!username) throw new BadRequestError('/body/username', 'missing required field');
-        if (username.length > 23) throw new BadRequestError('/body/username', 'username must be 23 characters or less');
-        if (email.length > 255) throw new BadRequestError('/body/email', 'email must be 255 characters or less');
-        if (!email.includes("@")) throw new BadRequestError('/body/email', 'email must contain "@"'); // TODO regexes
-
-        const user = await UserService.createUser(email, username);
-
-        return res.status(201).json(user);
-    } catch (error: any) {
+        logger.error("listUsers: error with status " + error.status);
         next(error);
     }
 }
 
 export async function updateUser(req: any, res: any, next: any) {
+    const id = Number(req.params.id);
     try {
-        const id = Number(req.params.id);
         const email = req.body?.email;
         const username = req.body?.username;
 
@@ -57,21 +44,27 @@ export async function updateUser(req: any, res: any, next: any) {
         if (email && !email.includes("@")) throw new BadRequestError('/body/email', 'email must contain "@"');
 
         const user = await UserService.updateUser(id, email, username);
-        return res.status(200).json(user);
+        
+        logger.info("updateUser: success. userId: " + user.id);
+        return res.status(200).json(user.getBasic());
     } catch (error: any) {
+        logger.error("updateUser: error with status " + error.status + ". userId: " + id);
         next(error);
     }
 }
 
 export async function deleteUser(req: any, res: any, next: any) {
+    const id = Number(req.params.id);
     try {
-        const id = Number(req.params.id);
-
         if (isNaN(id)) throw new BadRequestError('/id', 'id must be a number');
+        if (id == req.user.id) throw new ForbiddenError('Cannot delete currently authenticated user');
 
         await UserService.deleteUser(req.params.id);
+        
+        logger.info("deleteUser: success. userId: " + id);
         return res.status(204).json();
     } catch (error: any) {
+        logger.error("deleteUser: error with status " + error.status + ". userId: " + id);
         next(error);
     }
 }
